move consumer model inside catalog project
